Server/pipeline/image-sr/src/requirements.txt
Pillow
numpy
requests
google-cloud-storage
tensorflow-hub
google-cloud-pubsub
